requirements.txt
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
click==3.1
command-not-found==0.3
cryptography==41.0.3
dbus-python==1.2.18
decorator==5.1.1
distro==1.8.0
distro-info==1.1+ubuntu0.1
et-xmlfile==1.1.0
executing==1.2.0
hexlet==0.6.3
hexlet-code @ file:///home/ai_gelios/python-project-49/dist/hexlet_code-0.1.0-py3-none-any.whl#sha256=1cad9ef449a3987c344497760dfc5ac3574ee5aecce0fef189566bf44beb2a6a
httplib2==0.22.0
idna==2.5
importlib-metadata==6.8.0
ipython==8.14.0
jaraco.classes==3.3.0
jedi==0.18.2
jeepney==0.8.0
keyring==24.2.0
launchpadlib==1.11.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
matplotlib-inline==0.1.6
more-itertools==10.1.0
netifaces==0.11.0
oauthlib==3.2.2
openpyxl==3.1.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pprintpp==0.4.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
purl==0.8
pycparser==2.21
pyenv-win==3.1.1
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.8.0
pyparsing==3.1.1
python-apt==2.4.0+ubuntu2
python-i18n==0.3.0
PyYAML==3.11
requests==2.3.0
SecretStorage==3.3.3
six==1.16.0
stack-data==0.6.2
systemd-python==234
traitlets==5.9.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.21.1
wadllib==1.3.6
wcwidth==0.2.6
zipp==3.16.2
